refactor(dto): rename folder `schema` -> `dto`
