+++ src/Server.ts
@@ -2,7 +2,7 @@ import fastify from "fastify";
 import type { FastifyCookieOptions } from "@fastify/cookie";
 import { COOKIE_SECRET, CORS_WHITE_LIST, ENVIRONMENT, loggerConfig, swaggerConfig, swaggerUIConfig } from "@configs";
 import { apiPlugin, authPlugin } from "./plugins";
-import { wQuerySchema } from "@schemas/in";
+import { wQuerySchema } from "@dtos/in";
 import { kafkaAdmin } from "@services";
 import { wSetupHandler } from "@handlers";
