+++ src/dtos/in/cluster.dto.ts
@@ -1,4 +1,4 @@
-import { nullable } from "@schemas/common";
+import { nullable } from "src/dtos/common.schema";
 import s from "fluent-json-schema";
 
 export const clusterMutationSchema = s.object().prop("name", s.string().required()).prop("remarks", s.string().raw(nullable));
