+++ src/dtos/in/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from "./auth.dto";
+export * from "./cluster.dto";
+export * from "./script.dto";
+export * from "./sensor.dto";
