+++ src/handlers/user.handler.ts
@@ -1,6 +1,6 @@
 import { USER_NOT_FOUND } from "@constants";
 import { prisma } from "@repositories";
-import { FilterTemplateDto, KafkaBrokerDto, UserDto } from "@schemas/out";
+import { FilterTemplateDto, KafkaBrokerDto, UserDto } from "@dtos/out";
 import { FastifyReply, FastifyRequest } from "fastify";
 
 async function getUserById(request: FastifyRequest<{ Headers: { userId: string } }>, reply: FastifyReply): Result<UserDto> {
