--- src/schemas/in/index.ts
@@ -1,8 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from "./auth.schema";
-export * from "./cluster.schema";
-export * from "./script.schema";
-export * from "./sensor.schema";
