+++ tsconfig.json
@@ -32,11 +32,11 @@
             "@constants/*": ["src/constants/*"],
             "@services": ["src/services"],
             "@services/*": ["src/services/*"],
-            "@schemas/in": ["src/schemas/in"],
-            "@schemas/in/*": ["src/schemas/in/*"],
-            "@schemas/out": ["src/schemas/out"],
-            "@schemas/out/*": ["src/schemas/out/*"],
-            "@schemas/common": ["src/schemas/common.schema"],
+            "@dtos/in": ["src/dtos/in"],
+            "@dtos/in/*": ["src/dtos/in/*"],
+            "@dtos/out": ["src/dtos/out"],
+            "@dtos/out/*": ["src/dtos/out/*"],
+            "@dtos/common": ["src/dtos/common.dto"],
             "@configs": ["src/configs"],
             "@configs/*": ["src/configs/*"],
             "@repositories": ["src/prisma"],
